ad generator/requirements.txt
fastapi
uvicorn
pydantic

torch
transformers
accelerate
safetensors

diffusers
pillow
numpy
controlnet-aux
opencv-python-headless



openai
python-dotenv